### Commits

Cherry-picking 4 patches from upstream:

1. [bonding: add ESP offload features when slaves support](https://lore.kernel.org/all/20241105192721.584822-1-tariqt@nvidia.com/)
This patch fixes a perf issue in the bonding driver where ESP segmentation was not being offloaded to the slave devices.

2. [bonding: Correctly support GSO ESP offload](https://lore.kernel.org/netdev/20250127104147.759658-1-cratiu@nvidia.com/)
The fix in patch 1 is incomplete and this change fixes the ESP segmentation issue.
 
3. [xfrm_output: Force software GSO only in tunnel mode](https://lore.kernel.org/netdev/20250219105248.226962-1-cratiu@nvidia.com/)
Fix in xfrm_output to only trigger software GSO in crypto offload cases for already encapsulated packets in tunnel mode.

4. [xfrm: Support ESN context update to hardware for TX](https://lore.kernel.org/all/874f965d786606b0b4351c976f50271349f68b03.1734611621.git.leon@kernel.org/)
Previously xfrm_dev_state_advance_esn() was added for RX only in commit 50bd870a9e5cca9, but is needed for TX as well.

### Build

```
/mnt/ciq/sneginha/kernel-src-tree
  CLEAN   arch/x86/crypto
  CLEAN   arch/x86/entry/vdso
  CLEAN   arch/x86/kernel/cpu
  CLEAN   arch/x86/kernel
  CLEAN   arch/x86/purgatory
  CLEAN   arch/x86/realmode/rm
  CLEAN   arch/x86/lib
  CLEAN   certs
  CLEAN   crypto/asymmetric_keys

<SNIP>

[TIMER]{MRPROPER}: 19s
x86_64 architecture detected, copying config
'configs/kernel-5.14.0-x86_64.config' -> '.config'
Setting Local Version for build
CONFIG_LOCALVERSION="-sneginhal-fips-9-compliant_5.14.0-284.30.1-03b404b81bf2"
Making olddefconfig
  HOSTCC  scripts/basic/fixdep
  HOSTCC  scripts/kconfig/conf.o
  HOSTCC  scripts/kconfig/confdata.o

<SNIP>

  BTF [M] sound/usb/usx2y/snd-usb-usx2y.ko
  BTF [M] sound/virtio/virtio_snd.ko
  BTF [M] sound/xen/snd_xen_front.ko
  BTF [M] virt/lib/irqbypass.ko
[TIMER]{BUILD}: 2333s
Making Modules
  INSTALL /lib/modules/5.14.0-sneginhal-fips-9-compliant_5.14.0-284.30.1-03b404b81bf2+/kernel/arch/x86/crypto/blake2s-x86_64.ko
  STRIP   /lib/modules/5.14.0-sneginhal-fips-9-compliant_5.14.0-284.30.1-03b404b81bf2+/kernel/arch/x86/crypto/blake2s-x86_64.ko
  SIGN    /lib/modules/5.14.0-sneginhal-fips-9-compliant_5.14.0-284.30.1-03b404b81bf2+/kernel/arch/x86/crypto/blake2s-x86_64.ko
  INSTALL /lib/modules/5.14.0-sneginhal-fips-9-compliant_5.14.0-284.30.1-03b404b81bf2+/kernel/arch/x86/crypto/blowfish-x86_64.ko
  STRIP   /lib/modules/5.14.0-sneginhal-fips-9-compliant_5.14.0-284.30.1-03b404b81bf2+/kernel/arch/x86/crypto/blowfish-x86_64.ko
  SIGN    /lib/modules/5.14.0-sneginhal-fips-9-compliant_5.14.0-284.30.1-03b404b81bf2+/kernel/arch/x86/crypto/blowfish-x86_64.ko

<SNIP>

  STRIP   /lib/modules/5.14.0-sneginhal-fips-9-compliant_5.14.0-284.30.1-03b404b81bf2+/kernel/virt/lib/irqbypass.ko
  SIGN    /lib/modules/5.14.0-sneginhal-fips-9-compliant_5.14.0-284.30.1-03b404b81bf2+/kernel/virt/lib/irqbypass.ko
  DEPMOD  /lib/modules/5.14.0-sneginhal-fips-9-compliant_5.14.0-284.30.1-03b404b81bf2+
[TIMER]{MODULES}: 46s
Making Install
sh ./arch/x86/boot/install.sh \
        6.14.0-sneginhal-fips-9-compliant_5.14.0-284.30.1-03b404b81bf2+ arch/x86/boot/bzImage \
        System.map "/boot"
[TIMER]{INSTALL}: 25s
Checking kABI
kABI check passed
Setting Default Kernel to /boot/vmlinuz-5.14.0-sneginhal-fips-9-compliant_5.14.0-284.30.1-00adedc58a8e+ and Index to 1
Hopefully Grub2.0 took everything ... rebooting after time metrices
[TIMER]{MRPROPER}: 19s
[TIMER]{BUILD}: 2333s
[TIMER]{MODULES}: 46s
[TIMER]{INSTALL}: 25s
[TIMER]{TOTAL} 2431s
Rebooting in 10 seconds

```

### kABI Check

```
Checking kABI
kABI check passed
```

###  Reboot and Verify
```
Linux localhost 5.14.0-sneginhal-fips-9-compliant_5.14.0-284.30.1-00adedc58a8e+ #1 SMP PREEMPT_DYNAMIC Thu Mar 20 06:03:26 UTC 2025 x86_64 x86_64 x86_64 GNU/Linux
```

### Kernel Self Tests

[kselftest_output.txt](https://github.com/user-attachments/files/19374666/kselftest_output.txt)