@sneginha 2025-03-20
> Thanks for contributing to our kernel.
> 
> I appreciate following everything in the wiki for contribution guidelines.
> 
> I have a couple asks though: Could you amend these commit headers to include the full 40 character SHA (there is clearly a limitation in our tool and I've [created an issue to fix that](https://github.com/ctrliq/kernel-src-tree-tools/issues/7)) wiki updated as well: https://github.com/ctrliq/kernel-src-tree/wiki#tldr

Done.

> 
> * [bonding: Correctly support GSO ESP offload ](https://github.com/ctrliq/kernel-src-tree/pull/171/commits/c5daef8d5290dc6486cb74d94eea16b8a3388d87)
> * [xfrm_output: Force software GSO only in tunnel mode ](https://github.com/ctrliq/kernel-src-tree/pull/171/commits/54b97593983981dfa2bd2551249272997d82ab21)
> 
> Do you have testing that you can publicly share that validates that thes changes do what is expected?

Testing done:
1. Run iperf TCP performance tests
2. Review packet captures on the tx side
3. Ensure that you see large MSS sized packets going to the NIC.
4. Also, note the increase in network performance.